added user login redux logic
