--- client/src/context/resourcePost/resourcePostContext.js
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const resourcePostContext = createContext();
-
-export default resourcePostContext;
