--- client/src/context/types.js
@@ -1,4 +1,0 @@
-export const GET_N_POSTS = "GET_N_POSTS";
-export const SET_PAGE = "SET_PAGE";
-export const SET_POSTS = "SET_POSTS";
-export const SET_LOADING = "SET_LOADING"
